refactor: Reorganize plate reader module
