+++ lambda/src/data_hub_lambda/spectramax_plate_reader/__init__.py
@@ -1,0 +1,6 @@
+from data_hub_lambda.spectramax_plate_reader.parse_metadata import (
+    parse_metadata,  # noqa: F401
+)
+from data_hub_lambda.spectramax_plate_reader.process_file import (
+    process_file,  # noqa: F401
+)
