+++ lambda/src/data_hub_lambda/spectramax_plate_reader/parse_metadata.py
