--- lambda/src/data_hub_lambda/workflows/spectramax_id3_plate_reader/__init__.py
@@ -1,6 +1,0 @@
-from data_hub_lambda.workflows.spectramax_id3_plate_reader.parse_metadata import (
-    parse_metadata,  # noqa: F401
-)
-from data_hub_lambda.workflows.spectramax_id3_plate_reader.process_file import (
-    process_file,  # noqa: F401
-)
